:bug:fix: feed, comment, modal 컴포넌트 수정 #37
